refactor(live_edit): move HTML to separate file
Since we can reuse the HTML itself, let's put it in a separate file we
can include in live_edit.php and later live_create.php.
